+++ lib/td_bg/business_concepts/business_concepts.ex
@@ -304,9 +304,25 @@ defmodule TdBg.BusinessConcepts do
   end
 
   def update_business_concept_version_status(
+    %BusinessConceptVersion{} = business_concept_version,
+    %{status: "deprecated"} = attrs
+  ) do
+    result = do_update_business_concept_version_status(business_concept_version, attrs)
+    ConceptCache.delete(business_concept_version.business_concept_id)
+    result
+  end
+
+  def update_business_concept_version_status(
         %BusinessConceptVersion{} = business_concept_version,
         attrs
       ) do
+    do_update_business_concept_version_status(business_concept_version, attrs)
+  end
+
+  defp do_update_business_concept_version_status(
+         %BusinessConceptVersion{} = business_concept_version,
+         attrs
+       ) do
     result =
       business_concept_version
       |> BusinessConceptVersion.update_status_changeset(attrs)
@@ -569,7 +585,7 @@ defmodule TdBg.BusinessConcepts do
             },
             "business_concept:events"
           )
-
+          ConceptCache.delete(business_concept_id)
           # TODO: TD-1618 delete_search should be performed by a consumer of the event stream
           @search_service.delete_search(business_concept_version)
           {:ok, version}