+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - [TD-1946] Fixed error indexing concepts pertaining to soft-deleted domains
+- [TD-1687] Delete business concept from Concept Cache on bc first draft deletion and version deprecation
 
 ### Changed